requirements.txt
blinker==1.8.2
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.10.0
filelock==3.14.0
Flask==3.0.3
flask_sqlalchemy==3.0.3
fonttools==4.52.1
fsspec==2024.5.0
idna==2.10
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
opencv-python-headless==4.8.0.74
packaging==24.0
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
PyYAML==6.0.1
psycopg2-binary==2.9.7
requests==2.32.2
requests-toolbelt==1.0.0
roboflow==1.1.30
six==1.16.0
sympy==1.12
tbb==2021.12.0
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
typing_extensions==4.12.0
urllib3==2.2.1
Werkzeug==3.0.3
